requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.6
scikit-learn
pymongo-4.13.2

##-e . #refers to the setup.py file